Moved all layouts to their dedicated directories
